svg/painting/parsing/marker-shorthand.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: marker sets longhands</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#MarkerProperty" rel="help"/><h:meta content="marker supports the full paint grammar 'none | marker-ref'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/shorthand-testcommon.js"/><script><![CDATA[

test_shorthand_value('marker', 'none', {
  'marker-start': 'none',
  'marker-mid': 'none',
  'marker-end': 'none'
});

test_shorthand_value('marker', 'url("https://example.com/")', {
  'marker-start': 'url("https://example.com/")',
  'marker-mid': 'url("https://example.com/")',
  'marker-end': 'url("https://example.com/")'
});

  ]]></script></svg>
